Module Alpha_context.Sc_rollup_repr

This is a forward declaration to avoid circular dependencies. Use module Sc_rollup instead whenever possible. TODO : find a better way to resolve the circular dependency https://gitlab.com/tezos/tezos/-/issues/3147

module Address : Tezos_protocol_environment_015_PtLimaPt.S.HASH

type t = Address.t