Module Alpha_context.Tx_rollup_message

This module re-exports definitions from Tx_rollup_message_repr.

type deposit = {

sender : public_key_hash;

destination : Tx_rollup_l2_address.Indexable.value;

ticket_hash : Ticket_hash.t;

amount : Tx_rollup_l2_qty.t;

}

type t = private

| Batch of string

| Deposit of deposit

val make_batch : string -> t * int

make_batch batch creates a new Batch message to be added that can be added to an inbox, along with its size in bytes. See Tx_rollup_message_repr.size.

val make_deposit : public_key_hash -> Tx_rollup_l2_address.t Indexable.value -> Ticket_hash.t -> Tx_rollup_l2_qty.t -> t * int

make_deposit destination ticket_hash qty creates a new Deposit message to be added that can be added to an inbox, along with its size in bytes. See Tx_rollup_message_repr.size.

val encoding : t Tezos_protocol_environment_015_PtLimaPt.Data_encoding.t

val pp : Tezos_protocol_environment_015_PtLimaPt.Format.formatter -> t -> unit