Module Tezos_raw_protocol_015_PtLimaPt.Dal_apply

This modules handles all the validation/application/finalisation of any operation related to the DAL.

val validate_data_availability : Alpha_context.t -> Alpha_context.Dal.Endorsement.t -> unit Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult

validate_data_availability ctxt endorsement ensures the endorsement is valid and cannot prevent an operation containing endorsement to be refused on top of ctxt. If an Error _ is returned, the endorsement is not valid.

val apply_data_availability : Alpha_context.t -> Alpha_context.Dal.Endorsement.t -> endorser:Tezos_protocol_environment_015_PtLimaPt.Signature.Public_key_hash.t -> Alpha_context.t Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

apply_data_availability ctxt endorsement ~endorser applies endorsement into the ctxt assuming endorser issued those endorsements.

val validate_publish_slot_header : Alpha_context.t -> Alpha_context.Dal.Slot.t -> unit Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult

validate_publish_slot_header ctxt slot ensures that slot is valid and cannot prevent an operation containing slot to be refused on top of ctxt. If an Error _ is returned, the slot is not valid.

val apply_publish_slot_header : Alpha_context.t -> Alpha_context.Dal.Slot.t -> Alpha_context.t Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult

apply_publish_slot_header ctxt slot applies the publication of slot header slot on top of ctxt. Fails if the slot contains already a slot header.

val dal_finalisation : Alpha_context.t -> (Alpha_context.t * Alpha_context.Dal.Endorsement.t option) Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

dal_finalisation ctxt should be executed at block finalisation time. A set of slots available at level ctxt.current_level - lag is returned encapsulated into the endorsement data-structure.

lag is a parametric constant specific to the data-availability layer.