Module Dal_slot_repr.Page

A DAL slot is decomposed to a successive list of pages with fixed content size. The size is chosen so that it's possible to inject a page in a Tezos L1 operation if needed during the proof phase of a refutation game.

type content = Tezos_protocol_environment_015_PtLimaPt.Bytes.t

module Index : sig ... end

type t = {

slot_index : slot_index;

page_index : Index.t;

}

A page is identified by its slots index and by its own index in the list of pages of the slot.

val equal : t -> t -> bool

val encoding : t Tezos_protocol_environment_015_PtLimaPt.Data_encoding.t

val pp : Tezos_protocol_environment_015_PtLimaPt.Format.formatter -> t -> unit