Module Dal_slot_repr.Slots_history

This module provides an abstract data structure (type t) that represents a skip list used to store successive DAL slots confirmed on L1. There is one slot per cell in the skip list. The slots are sorted in increasing order by level, and by slot index, for the slots of the same level.

This module also defines a bounded history cache (type History_cache.t) that allows to remember recent values of a skip list of type t (indexed by the skip lists' hashes). This structure is meant to be maintained and used by the rollup node to produce refutation proofs involving DAL slot inputs.

type t

Abstract representation of a skip list specialized for confirmed slot headers.

val encoding : t Tezos_protocol_environment_015_PtLimaPt.Data_encoding.t

Encoding of the datatype.

val genesis : t

First cell of this skip list.

module History_cache : Bounded_history_repr.S

The History_cache.t structure is basically a bounded lookup table of t skip lists. (See Bounded_history_repr.S). In the L1 layer, the capacity (bound) is set to zero (nothing is remembered). By contrast, the rollup node uses a history cache with a (sufficiently) large capacity to participate in all potential refutation games occurring during the challenge period. Indeed, the successive recent skip-lists stored in the cache are needed to produce proofs involving slots' pages.

val add_confirmed_slots : t -> History_cache.t -> slot list -> (t * History_cache.t) Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult

add_confirmed_slots hist cache slots updates the given structure hist with the list of slots. The given cache is also updated to add successive values of cell to it.

val add_confirmed_slots_no_cache : t -> slot list -> t Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult

add_confirmed_slots_no_cache cell slots same as add_confirmed_slots, but no cache is updated.

val equal : t -> t -> bool

equal a b returns true iff a is equal to b.