Module Tezos_raw_protocol_015_PtLimaPt.Delegate_services

This module defines RPC services to access the information associated to delegates (who they are, their delegators, their different kinds of balances, their activity, etc.).

type Tezos_protocol_environment_015_PtLimaPt.Error_monad.error +=

| Not_registered of Tezos_protocol_environment_015_PtLimaPt.Signature.Public_key_hash.t

val list : 'a Tezos_protocol_environment_015_PtLimaPt.RPC_context.simple -> 'a -> ?active:bool -> ?inactive:bool -> ?with_minimal_stake:bool -> ?without_minimal_stake:bool -> unit -> Tezos_protocol_environment_015_PtLimaPt.Signature.Public_key_hash.t list Tezos_protocol_environment_015_PtLimaPt.Error_monad.shell_tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

type info = {

full_balance : Alpha_context.Tez.t;
(*
Balance + Frozen balance
*)

current_frozen_deposits : Alpha_context.Tez.t;

frozen_deposits : Alpha_context.Tez.t;

staking_balance : Alpha_context.Tez.t;

frozen_deposits_limit : Alpha_context.Tez.t option;

delegated_contracts : Alpha_context.Contract.t list;

delegated_balance : Alpha_context.Tez.t;

deactivated : bool;

grace_period : Alpha_context.Cycle.t;

voting_info : Alpha_context.Vote.delegate_info;

active_consensus_key : Tezos_protocol_environment_015_PtLimaPt.Signature.Public_key_hash.t;

pending_consensus_keys : (Alpha_context.Cycle.t * Tezos_protocol_environment_015_PtLimaPt.Signature.Public_key_hash.t) list;

}

val info_encoding : info Tezos_protocol_environment_015_PtLimaPt.Data_encoding.t

val info : 'a Tezos_protocol_environment_015_PtLimaPt.RPC_context.simple -> 'a -> Tezos_protocol_environment_015_PtLimaPt.Signature.Public_key_hash.t -> info Tezos_protocol_environment_015_PtLimaPt.Error_monad.shell_tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

val full_balance : 'a Tezos_protocol_environment_015_PtLimaPt.RPC_context.simple -> 'a -> Tezos_protocol_environment_015_PtLimaPt.Signature.Public_key_hash.t -> Alpha_context.Tez.t Tezos_protocol_environment_015_PtLimaPt.Error_monad.shell_tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

val current_frozen_deposits : 'a Tezos_protocol_environment_015_PtLimaPt.RPC_context.simple -> 'a -> Tezos_protocol_environment_015_PtLimaPt.Signature.Public_key_hash.t -> Alpha_context.Tez.t Tezos_protocol_environment_015_PtLimaPt.Error_monad.shell_tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

val frozen_deposits : 'a Tezos_protocol_environment_015_PtLimaPt.RPC_context.simple -> 'a -> Tezos_protocol_environment_015_PtLimaPt.Signature.Public_key_hash.t -> Alpha_context.Tez.t Tezos_protocol_environment_015_PtLimaPt.Error_monad.shell_tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

val staking_balance : 'a Tezos_protocol_environment_015_PtLimaPt.RPC_context.simple -> 'a -> Tezos_protocol_environment_015_PtLimaPt.Signature.Public_key_hash.t -> Alpha_context.Tez.t Tezos_protocol_environment_015_PtLimaPt.Error_monad.shell_tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

val frozen_deposits_limit : 'a Tezos_protocol_environment_015_PtLimaPt.RPC_context.simple -> 'a -> Tezos_protocol_environment_015_PtLimaPt.Signature.Public_key_hash.t -> Alpha_context.Tez.t option Tezos_protocol_environment_015_PtLimaPt.Error_monad.shell_tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

val delegated_contracts : 'a Tezos_protocol_environment_015_PtLimaPt.RPC_context.simple -> 'a -> Tezos_protocol_environment_015_PtLimaPt.Signature.Public_key_hash.t -> Alpha_context.Contract.t list Tezos_protocol_environment_015_PtLimaPt.Error_monad.shell_tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

val delegated_balance : 'a Tezos_protocol_environment_015_PtLimaPt.RPC_context.simple -> 'a -> Tezos_protocol_environment_015_PtLimaPt.Signature.Public_key_hash.t -> Alpha_context.Tez.t Tezos_protocol_environment_015_PtLimaPt.Error_monad.shell_tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

val deactivated : 'a Tezos_protocol_environment_015_PtLimaPt.RPC_context.simple -> 'a -> Tezos_protocol_environment_015_PtLimaPt.Signature.Public_key_hash.t -> bool Tezos_protocol_environment_015_PtLimaPt.Error_monad.shell_tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

val grace_period : 'a Tezos_protocol_environment_015_PtLimaPt.RPC_context.simple -> 'a -> Tezos_protocol_environment_015_PtLimaPt.Signature.Public_key_hash.t -> Alpha_context.Cycle.t Tezos_protocol_environment_015_PtLimaPt.Error_monad.shell_tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

val voting_power : 'a Tezos_protocol_environment_015_PtLimaPt.RPC_context.simple -> 'a -> Alpha_context.public_key_hash -> int64 Tezos_protocol_environment_015_PtLimaPt.Error_monad.shell_tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

val voting_info : 'a Tezos_protocol_environment_015_PtLimaPt.RPC_context.simple -> 'a -> Alpha_context.public_key_hash -> Alpha_context.Vote.delegate_info Tezos_protocol_environment_015_PtLimaPt.Error_monad.shell_tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

val consensus_key : 'a Tezos_protocol_environment_015_PtLimaPt.RPC_context.simple -> 'a -> Tezos_protocol_environment_015_PtLimaPt.Signature.Public_key_hash.t -> (Tezos_protocol_environment_015_PtLimaPt.Signature.Public_key_hash.t * (Alpha_context.Cycle.t * Tezos_protocol_environment_015_PtLimaPt.Signature.Public_key_hash.t) list) Tezos_protocol_environment_015_PtLimaPt.Error_monad.shell_tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

val participation : 'a Tezos_protocol_environment_015_PtLimaPt.RPC_context.simple -> 'a -> Alpha_context.public_key_hash -> Alpha_context.Delegate.participation_info Tezos_protocol_environment_015_PtLimaPt.Error_monad.shell_tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

val register : unit -> unit