Module Tezos_raw_protocol_015_PtLimaPt.Delegate_slashed_deposits_storage

This module maintains the storage related to slashing of delegates for double signing. In particular, it is responsible for maintaining the Storage.Slashed_deposits table.

val already_slashed_for_double_baking : Raw_context.t -> Tezos_protocol_environment_015_PtLimaPt.Signature.Public_key_hash.t -> Level_repr.t -> bool Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

Returns true if the given delegate has already been slashed for double baking for the given level.

val already_slashed_for_double_endorsing : Raw_context.t -> Tezos_protocol_environment_015_PtLimaPt.Signature.Public_key_hash.t -> Level_repr.t -> bool Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

Returns true if the given delegate has already been slashed for double preendorsing or double endorsing for the given level.

val punish_double_endorsing : Raw_context.t -> Tezos_protocol_environment_015_PtLimaPt.Signature.Public_key_hash.t -> Level_repr.t -> (Raw_context.t * Tez_repr.t * Receipt_repr.balance_updates) Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

Burn some frozen deposit for a delegate at a given level and record in the context that the given delegate has now been slashed for double endorsing for the given level.

Returns the burned amount.

Fails with Unrequired_denunciation if the given delegate has already been slashed for double endorsing for the given level.

val punish_double_baking : Raw_context.t -> Tezos_protocol_environment_015_PtLimaPt.Signature.Public_key_hash.t -> Level_repr.t -> (Raw_context.t * Tez_repr.t * Receipt_repr.balance_updates) Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

Burn some frozen deposit for a delegate at a given level and record in the context that the given delegate has now been slashed for double baking for the given level.

Returns the burned amount.

Fails with Unrequired_denunciation if the given delegate has already been slashed for double baking for the given level.

val clear_outdated_slashed_deposits : Raw_context.t -> new_cycle:Cycle_repr.t -> Raw_context.t Tezos_protocol_environment_015_PtLimaPt.Lwt.t