Module type Lazy_storage_kind.ID

Lazy storage ids are kept as abstract as possible to avoid mixing them up.

Behind the scene they are Z.ts but, within the protocol, only parse_data/ unparse_data are allowed convert from/to it.

Temporary ids may be used to pass values between contracts that won't be kept longer than the lifetime of the operation. Behind the scene, temporary ids are negative Z.ts.

type t

val compare : t -> t -> int

val encoding : t Tezos_protocol_environment_015_PtLimaPt.Data_encoding.t

val rpc_arg : t Tezos_protocol_environment_015_PtLimaPt.RPC_arg.arg

val init : t

Initial value for ids: zero.

val parse_z : Tezos_protocol_environment_015_PtLimaPt.Z.t -> t

In the protocol, to be used in parse_data only

val unparse_to_z : t -> Tezos_protocol_environment_015_PtLimaPt.Z.t

In the protocol, to be used in unparse_data only

val next : t -> t

val is_temp : t -> bool

val of_legacy_USE_ONLY_IN_Legacy_big_map_diff : Tezos_protocol_environment_015_PtLimaPt.Z.t -> t

val to_legacy_USE_ONLY_IN_Legacy_big_map_diff : t -> Tezos_protocol_environment_015_PtLimaPt.Z.t

include Path_encoding.S with type t := t

val to_path : t -> string list -> string list

to_path t postfix returns the context path name for t postfixed with postfix

val of_path : string list -> t option

of_path path parses path as a context path name for t

val path_length : int

Directory levels of the path encoding of t