Module Tezos_raw_protocol_015_PtLimaPt.Nonce_storage

This module provides types and functions to manipulate nonces.

A nonce is a byte sequence of fixed length, which is supposed to be random and used only once, provided by a block producer and used to generate a random seed (see Seed_repr).

type t = Seed_repr.nonce

type nonce = t

type Tezos_protocol_environment_015_PtLimaPt.Error_monad.error +=

| Too_late_revelation

| Too_early_revelation

| Already_revealed_nonce

| Inconsistent_nonce

val encoding : nonce Tezos_protocol_environment_015_PtLimaPt.Data_encoding.t

type unrevealed = Storage.Seed.unrevealed_nonce = {

nonce_hash : Nonce_hash.t;

delegate : Tezos_protocol_environment_015_PtLimaPt.Signature.Public_key_hash.t;

}

type status =

| Unrevealed of unrevealed

| Revealed of Seed_repr.nonce

val get : Raw_context.t -> Level_repr.t -> status Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

type nonce_presence =

| No_nonce_expected

| Nonce_expected of status

val check : Raw_context.t -> Level_repr.t -> nonce_presence Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

val record_hash : Raw_context.t -> unrevealed -> Raw_context.t Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

val check_unrevealed : Raw_context.t -> Level_repr.t -> nonce -> unit Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

Checks that a nonce revelation operation can be safely applied.

returns
Error Too_early_revelation if the current cycle is the cycle 0 or if the previous cycle is lesser than the cycle of the input level.

returns
Error Too_late_revelation if the previous cycle is greater than the cycle of the input level. This error is also returned if the current level cycle position is greater or equal to the nonce revelation threshold.

returns
Error Already_revealed_nonce if a nonce is already revealed in the context for the input level.

returns
Error Inconsistent_nonce if the hash of the input nonce does not correspond to the nonce recover from the context for the given level.

val reveal : Raw_context.t -> Level_repr.t -> nonce -> Raw_context.t Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

val of_bytes : bytes -> nonce Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult

val hash : nonce -> Nonce_hash.t

val check_hash : nonce -> Nonce_hash.t -> bool