Module Raw_context.Dal

val record_available_shards : t -> Dal_endorsement_repr.t -> int list -> t

record_available_shards ctxt slots shards records that the list of shards shards were declared available. The function assumes that a shard belongs to the interval 0; number_of_shards - 1. Otherwise, for each shard outside this interval, it is a no-op.

val register_slot : t -> Dal_slot_repr.t -> (t * bool) Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult

register_slot ctxt slot returns a new context where the new candidate slot have been taken into account. Returns Some (ctxt,updated) where updated=true if the candidate is registered. Some (ctxt,false) if another candidate was already registered previously. Returns an error if the slot is invalid.

val candidates : t -> Dal_slot_repr.t list

candidates ctxt returns the current list of slot for which there is at least one candidate.

val is_slot_available : t -> Dal_slot_repr.Index.t -> bool

is_slot_available ctxt slot_index returns true if the slot_index is declared available by the protocol. false otherwise. If the index is out of the interval 0;number_of_slots - 1, returns false.

val shards : t -> endorser:Tezos_protocol_environment_015_PtLimaPt.Signature.Public_key_hash.t -> int list

shards ctxt ~endorser returns the shard assignment for the endorser for the current level.