Module Sc_rollup_inbox_repr.Make_hashing_scheme

This validation is based on a standardized Merkelization scheme. The definition of this scheme is independent from the exact data model of the context but it depends on the Tree arity and internal hashing scheme.

We provide a functor that takes a Context.TREE module from any context, checks that the assumptions made about tree's arity and hashing scheme are valid, and returns a standard compliant implementation of the Merkelized_operations.

Parameters

module P : P

Signature

type tree = P.tree

The type for the Merkle trees used in this module.

type inbox_context = P.t

The context used by the trees.

val hash_level_tree : tree -> Hash.t

Standard hashing function used for trees in this module.

val new_level_tree : inbox_context -> Raw_level_repr.t -> tree Tezos_protocol_environment_015_PtLimaPt.Lwt.t

Initialise a new level. new_level_tree ctxt level is a merkle tree with no messages yet, but has the level stored so we can check that in proofs.

val add_messages : inbox_context -> History.t -> t -> Raw_level_repr.t -> Sc_rollup_inbox_message_repr.serialized list -> tree option -> (tree * History.t * t) Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

add_messages ctxt history inbox level payloads level_tree inserts a list of payloads as new messages in the level_tree of the current level of the inbox. This function returns the new level tree as well as updated inbox and history.

If the inbox's level is older than level, the inbox is updated so that the level trees of the levels older than level are archived. To archive a level_tree for a given level, we push it at the end of the history and update the witness of this history in the inbox. The inbox's level tree for the current level is emptied to insert the payloads in a fresh level_tree for level.

This function fails if level is older than inbox's level.

val add_messages_no_history : inbox_context -> t -> Raw_level_repr.t -> Sc_rollup_inbox_message_repr.serialized list -> tree option -> (tree * t, Tezos_protocol_environment_015_PtLimaPt.Error_monad.error Tezos_protocol_environment_015_PtLimaPt.Error_monad.trace) Tezos_protocol_environment_015_PtLimaPt.Pervasives.result Tezos_protocol_environment_015_PtLimaPt.Lwt.t

add_messages_no_history ctxt inbox level payloads level_tree behaves as add_external_messages except that it does not remember the inbox history.

val get_message_payload : tree -> Tezos_protocol_environment_015_PtLimaPt.Z.t -> Sc_rollup_inbox_message_repr.serialized option Tezos_protocol_environment_015_PtLimaPt.Lwt.t

get_message_payload level_tree idx returns Some payload if the level_tree has more than idx messages, and payload is at position idx. Returns None otherwise.

val form_history_proof : inbox_context -> History.t -> t -> tree option -> (History.t * history_proof) Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

form_history_proof ctxt history inbox level_tree creates the skip list structure that includes the current inbox level, while also updating the history and making sure the level_tree has been committed to the ctxt.

This is used in archive_if_needed to produce the old_levels_messages value for the next level of the inbox. It is also needed if you want to produce a fully-up-to-date skip list for proof production. Just taking the skip list stored in the inbox at old_levels_messages will not include the current level (and that current level could be quite far back in terms of blocks if the inbox hasn't been added to for a while).

val take_snapshot : t -> history_proof

This is similar to form_history_proof except that it is just to be used on the protocol side because it doesn't ensure the history is remembered or the trees are committed in the context. Used at the beginning of a refutation game to create the snapshot against which proofs in that game must be valid.

This will however produce a history_proof with exactly the same hash as the one produced by form_history_proof, run on a node with a complete inbox_context.

type inclusion_proof

Given a inbox A at some level L and another inbox B at some level L' >= L, an inclusion_proof guarantees that A is an older version of B.

To be more precise, an inclusion_proof guarantees that the previous levels level_trees of A are included in the previous levels level_trees of B. The current level_tree of A and B are not considered.

The size of this proof is O(log_basis (L' - L)).

val inclusion_proof_encoding : inclusion_proof Tezos_protocol_environment_015_PtLimaPt.Data_encoding.t

val pp_inclusion_proof : Tezos_protocol_environment_015_PtLimaPt.Format.formatter -> inclusion_proof -> unit

val number_of_proof_steps : inclusion_proof -> int

number_of_proof_steps proof returns the length of proof.

val verify_inclusion_proof : inclusion_proof -> history_proof -> history_proof -> bool

verify_inclusion_proof proof a b returns true iff proof is a minimal and valid proof that a is included in b.

type proof

An inbox proof has three parameters:

the starting_point, of type Raw_level_repr.t * Z.t, specifying a location in the inbox ;

the message, of type Sc_rollup_PVM_sig.input option ;

and a reference snapshot inbox.

A valid inbox proof implies the following semantics: beginning at starting_point and reading forward through snapshot, the first message you reach will be message.

Usually this is fairly simple because there will actually be a message at the location specified by starting_point. But in some cases starting_point is past the last message within a level, and then the inbox proof must prove that and also provide another proof about the message at the beginning of the next non-empty level.

val pp_proof : Tezos_protocol_environment_015_PtLimaPt.Format.formatter -> proof -> unit

val to_serialized_proof : proof -> serialized_proof

val of_serialized_proof : serialized_proof -> proof option

val verify_proof : (Raw_level_repr.t * Tezos_protocol_environment_015_PtLimaPt.Z.t) -> history_proof -> proof -> Sc_rollup_PVM_sig.inbox_message option Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

See the docstring for the proof type for details of proof semantics.

verify_proof starting_point inbox proof will return the third parameter of the proof, message, iff the proof is valid.

val produce_proof : inbox_context -> History.t -> history_proof -> (Raw_level_repr.t * Tezos_protocol_environment_015_PtLimaPt.Z.t) -> (proof * Sc_rollup_PVM_sig.inbox_message option) Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

produce_proof ctxt history inbox (level, counter) creates an inbox proof proving the first message after the index counter at location level. This will fail if the ctxt given doesn't have sufficient data (it needs to be run on an inbox_context with the full history).

val empty : inbox_context -> Sc_rollup_repr.t -> Raw_level_repr.t -> t Tezos_protocol_environment_015_PtLimaPt.Lwt.t

empty ctxt level is an inbox started at some given level with no message at all.

module Internal_for_tests : sig ... end