Module Tezos_raw_protocol_015_PtLimaPt.Sc_rollup_inbox_storage

val inbox : Raw_context.t -> Sc_rollup_repr.t -> (Sc_rollup_inbox_repr.t * Raw_context.t) Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

inbox context rollup returns the current state of the inbox.

val add_external_messages : Raw_context.t -> Sc_rollup_repr.t -> string list -> (Sc_rollup_inbox_repr.t * Tezos_protocol_environment_015_PtLimaPt.Z.t * Raw_context.t) Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

add_external_messages context rollup msg adds msg to rollup's inbox.

This function returns the updated context as well as the size diff.

May fail with:

Sc_rollup_max_number_of_available_messages if inbox is full

Sc_rollup_max_number_of_messages_reached_for_commitment_period if the number of messages pushed during commitment period is too high

val add_internal_message : Raw_context.t -> Sc_rollup_repr.t -> payload:Script_repr.expr -> sender:Contract_hash.t -> source:Tezos_protocol_environment_015_PtLimaPt.Signature.public_key_hash -> (Sc_rollup_inbox_repr.t * Tezos_protocol_environment_015_PtLimaPt.Z.t * Raw_context.t) Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

add_internal_message context rollup ~payload ~sender ~source adds the internal message of payload, sender, and source to rollup's inbox.

See add_external_messages for returned values and failures.