Module Tezos_raw_protocol_015_PtLimaPt.Sc_rollup_outbox_storage

A module for managing state concerning a rollup's outbox.

val record_applied_message : Raw_context.t -> Sc_rollup_repr.t -> Raw_level_repr.t -> message_index:int -> (Tezos_protocol_environment_015_PtLimaPt.Z.t * Raw_context.t) Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

record_applied_message ctxt rollup level ~message_index marks the message in the outbox of rollup rollup at level level and position message_index as processed. Returns the size diff resulting from adding an entry. The size diff may be 0 if an entry already exists, or negative if an index is replaced with a new level.

An attempt to apply an old level that has already been replaced fails with an Sc_rollup_outbox_level_expired error.

In case a message has already been applied for the given level and message index, the function fails with an Sc_rollup_outbox_message_already_applied error.