Module Tezos_raw_protocol_015_PtLimaPt.Sc_rollup_storage

val originate : Raw_context.t -> kind:Sc_rollups.Kind.t -> boot_sector:string -> parameters_ty:Script_repr.lazy_expr -> genesis_commitment:Sc_rollup_commitment_repr.t -> (Sc_rollup_repr.Address.t * Tezos_protocol_environment_015_PtLimaPt.Z.t * Sc_rollup_commitment_repr.Hash.t * Raw_context.t) Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

originate context ~kind ~boot_sector produces an address a for a smart contract rollup using the origination nonce found in context. This function also initializes the storage with a new entry indexed by a to remember the kind of the rollup at address a and also to remember its boot_sector.

Also returns the number of allocated bytes.

val kind : Raw_context.t -> Sc_rollup_repr.t -> (Raw_context.t * Sc_rollups.Kind.t) Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

kind context address returns the kind of the given rollup address iff address is an existing rollup. Fails with an Sc_rollup_does_not_exist error in case the rollup does not exist.

val list_unaccounted : Raw_context.t -> Sc_rollup_repr.t list Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

val genesis_info : Raw_context.t -> Sc_rollup_repr.t -> (Raw_context.t * Sc_rollup_commitment_repr.genesis_info) Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

genesis_info ctxt sc_rollup returns the level at which a sc_rollup was originated, and its genesis commitment hash.

val get_boot_sector : Raw_context.t -> Sc_rollup_repr.t -> (Raw_context.t * string) Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

get_boot_sector ctxt sc_rollup retrieves the boot sector for sc_rollup.

val parameters_type : Raw_context.t -> Sc_rollup_repr.t -> (Script_repr.lazy_expr option * Raw_context.t) Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

parameters_type ctxt rollup returns the registered type of a rollup. Returns None in case there is no registered type for the rollup.

module Dal_slot : sig ... end