Module Sc_rollup_storage.Dal_slot

val subscribe : Raw_context.t -> Sc_rollup_repr.t -> slot_index:Dal_slot_repr.Index.t -> (Dal_slot_repr.Index.t * Raw_level_repr.t * Raw_context.t) Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

subscribe ctxt rollup slot_index marks the rollup as subscribed to slot_index at the level indicated by Raw_context.current_level ctxt.

May fail with:

Sc_rollup_does_not_exist if rollup does not exist

Dal_subscribe_rollup_invalid_slot_index of {given=slot_index; maximum} if the slot_index is either negative or above maximum, which is the maximum slot index (inclusive) allowed

Dal_errors_repr.Dal_rollup_already_registered_to_slot (rollup, slot_index) if rollup is already subscribed to slot_index

val subscribed_slot_indices : Raw_context.t -> Sc_rollup_repr.t -> Raw_level_repr.t -> Dal_slot_repr.Index.t list Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

subscribed_slot_indices ctxt rollup level returns the slots to which rollup was subscribed at level level.

May fail with:

Sc_rollup_does_not_exist if rollup does not exist

Dal_errors_repr.Dal_requested_subscriptions_at_future_level (current_level, level) if level is above the current elvel, i.e. current = Raw_context.current_level ctxt and level > current