Module Sc_rollup_wasm.V2_0_0

This module provides Proof-Generating Virtual Machine (PVM) running WebAssembly (version 2.0.0).

module type S = sig ... end

module type P = sig ... end

module type Make_wasm = module type of Tezos_protocol_environment_015_PtLimaPt.Wasm_2_0_0.Make

module Make (Lib_scoru_Wasm : Make_wasm) (Context : P) : S with type context = Context.Tree.t and type state = Context.tree and type proof = Context.proof

Build a WebAssembly PVM using the given proof-supporting context.

module Protocol_implementation : S with type context = Tezos_protocol_environment_015_PtLimaPt.Context.t and type state = Tezos_protocol_environment_015_PtLimaPt.Context.tree and type proof = Tezos_protocol_environment_015_PtLimaPt.Context.Proof.tree Tezos_protocol_environment_015_PtLimaPt.Context.Proof.t

This PVM is used for verification in the Protocol. produce_proof always returns None.

val reference_initial_state_hash : Sc_rollup_repr.State_hash.t

This is the state hash of reference that both the prover of the node and the verifier of the protocol Protocol_implementation have to agree on (if they do, it means they are using the same tree structure).