Module Tezos_raw_protocol_015_PtLimaPt.Script_typed_ir_size

This module provides overapproximation of memory footprint for Michelson-related values.

These overapproximations are used by the cache to evaluate its own memory footprint and enforce declared limit over its size.

val value_size : ('a, _) Script_typed_ir.ty -> 'a -> Cache_memory_helpers.nodes_and_size

value_size ty v returns an overapproximation of the size of the in-memory representation of v of type ty.

val lambda_size : ('a, 'b) Script_typed_ir.lambda -> Cache_memory_helpers.nodes_and_size

lambda_size l returns an overapproximation of the size of the internal IR for the Michelson lambda abstraction l.

val node_size : Script_repr.node -> Cache_memory_helpers.nodes_and_size

node_size root returns the size of the in-memory representation of root in bytes. This is an over-approximation of the memory actually consumed by root since no sharing is taken into account.

val (++) : Cache_memory_helpers.nodes_and_size -> Cache_memory_helpers.nodes_and_size -> Cache_memory_helpers.nodes_and_size

Pointwise addition (reexport from Cache_memory_helpers)

val zero : Cache_memory_helpers.nodes_and_size

Zero vector (reexport from Cache_memory_helpers)