Module Slot_repr.Range

Slot ranges

type t

An ordered range of slots, in increasing order.

Constructor

val create : min:int -> count:int -> t Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult

create ~min ~count creates a full slot range starting at min, of size count, i.e, min, min + count - 1.

create errors if

min < 0

count < 1

min + count - 1 > max_value

Iterators

val fold : ('a -> slot -> 'a) -> 'a -> t -> 'a

fold f acc range folds f over the values of range, in increasing order.

val fold_es : ('a -> slot -> 'a Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t) -> 'a -> t -> 'a Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

fold_es f acc range folds f over the values of range, in increasing order.

val rev_fold_es : ('a -> slot -> 'a Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t) -> 'a -> t -> 'a Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

rev_fold_es f acc range folds f over the values of range, in decreasing order.