Module Storage.Dal

module Slot_headers : Storage_sigs.Non_iterable_indexed_data_storage with type t = Raw_context.t and type key = Raw_level_repr.t and type value = Dal_slot_repr.slot list

This is a temporary storage for slot header proposed onto the L1.

module Slots_history : Storage_sigs.Single_data_storage with type t := Raw_context.t and type value = Dal_slot_repr.Slots_history.t

This is a permanent storage for slot header confirmed by the L1.