Module Storage.Liquidity_baking

module Toggle_ema : Storage_sigs.Single_data_storage with type t := Raw_context.t and type value = Tezos_protocol_environment_015_PtLimaPt.Int32.t

Exponential moving average (ema) of flags set in protocol_data.contents. The liquidity baking subsidy is not sent to the CPMM if this EMA is above the threshold set in constants. *

module Cpmm_address : Storage_sigs.Single_data_storage with type t := Raw_context.t and type value = Contract_hash.t

Constant product market maker contract that receives liquidity baking subsidy. *