Module Storage.Sc_rollup

module PVM_kind : Storage_sigs.Non_iterable_indexed_carbonated_data_storage with type key = Sc_rollup_repr.t and type value = Sc_rollups.Kind.t and type t := Raw_context.t

Smart contract rollup.

module Boot_sector : Storage_sigs.Non_iterable_indexed_carbonated_data_storage with type key = Sc_rollup_repr.t and type value = string and type t := Raw_context.t

module Parameters_type : Storage_sigs.Non_iterable_indexed_carbonated_data_storage with type key = Sc_rollup_repr.t and type value = Script_repr.lazy_expr and type t := Raw_context.t

module Genesis_info : Storage_sigs.Non_iterable_indexed_carbonated_data_storage with type key = Sc_rollup_repr.t and type value = Sc_rollup_commitment_repr.genesis_info and type t := Raw_context.t

module Inbox : Storage_sigs.Non_iterable_indexed_carbonated_data_storage with type key = Sc_rollup_repr.t and type value = Sc_rollup_inbox_repr.t and type t := Raw_context.t

module Last_cemented_commitment : Storage_sigs.Non_iterable_indexed_carbonated_data_storage with type key = Sc_rollup_repr.t and type value = Sc_rollup_commitment_repr.Hash.t and type t := Raw_context.t

module Stakers : Storage_sigs.Non_iterable_indexed_carbonated_data_storage with type key = Tezos_protocol_environment_015_PtLimaPt.Signature.Public_key_hash.t and type value = Sc_rollup_commitment_repr.Hash.t and type t = Raw_context.t * Sc_rollup_repr.t

val stakers : Raw_context.t -> Sc_rollup_repr.t -> (Raw_context.t * (Tezos_protocol_environment_015_PtLimaPt.Signature.Public_key_hash.t * Sc_rollup_commitment_repr.Hash.t) list) Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult Tezos_protocol_environment_015_PtLimaPt.Lwt.t

stakers ctxt rollup returns all the stakers over rollup with their related commitment.

module Staker_count : Storage_sigs.Non_iterable_indexed_carbonated_data_storage with type key = Sc_rollup_repr.t and type value = int32 and type t := Raw_context.t

Cache: This should always be the number of entries in Stakers.

module Commitments : Storage_sigs.Non_iterable_indexed_carbonated_data_storage with type key = Sc_rollup_commitment_repr.Hash.t and type value = Sc_rollup_commitment_repr.t and type t = Raw_context.t * Sc_rollup_repr.t

module Commitment_stake_count : Storage_sigs.Non_iterable_indexed_carbonated_data_storage with type key = Sc_rollup_commitment_repr.Hash.t and type value = int32 and type t = Raw_context.t * Sc_rollup_repr.t

Cache: This should always be the number of stakers that are directly or indirectly staked on this commitment.

module Commitment_added : Storage_sigs.Non_iterable_indexed_carbonated_data_storage with type key = Sc_rollup_commitment_repr.Hash.t and type value = Raw_level_repr.t and type t = Raw_context.t * Sc_rollup_repr.t

module Game : Storage_sigs.Non_iterable_indexed_carbonated_data_storage with type key = Sc_rollup_game_repr.Index.t and type value = Sc_rollup_game_repr.t and type t = Raw_context.t * Sc_rollup_repr.t

Refutation games are indexed by the rollup and the pair of competing stakers. The staker pair should always be in lexical order to ensure that games are not duplicated.

module Game_timeout : Storage_sigs.Non_iterable_indexed_carbonated_data_storage with type key = Sc_rollup_game_repr.Index.t and type value = Sc_rollup_game_repr.timeout and type t = Raw_context.t * Sc_rollup_repr.t

Game_timeout stores the block level at which the staker whose turn it is to move will (become vulnerable to) timeout. The staker pair should always be in lexical order to ensure that this value is not duplicated.

module Opponent : Storage_sigs.Non_iterable_indexed_carbonated_data_storage with type key = Tezos_protocol_environment_015_PtLimaPt.Signature.Public_key_hash.t and type value = Sc_rollup_repr.Staker.t and type t = Raw_context.t * Sc_rollup_repr.t

Opponent stores the current opponent of the staker. This is mainly used to enforce the requirement that each staker should only play one refutation game at a time. It will also be useful for searching for current game by staker.

module Applied_outbox_messages : Storage_sigs.Non_iterable_indexed_carbonated_data_storage with type t = Raw_context.t * Sc_rollup_repr.t and type key = int32 and type value = Raw_level_repr.t * Bitset.t

A carbonated storage for keeping track of applied outbox messages for a a SCORU.

module Slot_subscriptions : Storage_sigs.Indexed_data_storage with type t = Raw_context.t * Sc_rollup_repr.t and type key = Raw_level_repr.t and type value = Bitset.t

An indexed data storage for keeping track of dal slots to which a rollup is subscribed to a given level.