Module Storage.Tenderbake

module First_level_of_protocol : Storage_sigs.Single_data_storage with type t := Raw_context.t and type value = Raw_level_repr.t

First_level_of_protocol stores the level of the first block of this protocol.

module Endorsement_branch : Storage_sigs.Single_data_storage with type value = Tezos_protocol_environment_015_PtLimaPt.Block_hash.t * Block_payload_hash.t and type t := Raw_context.t

Endorsement_branch stores a single value composed of the grandparent hash and the predecessor's payload (computed with the grandparent hash) used to verify the validity of endorsements.

module Grand_parent_branch : Storage_sigs.Single_data_storage with type value = Tezos_protocol_environment_015_PtLimaPt.Block_hash.t * Block_payload_hash.t and type t := Raw_context.t

Grand_parent_branch stores a single value composed of the great-grand parent hash and the grand parent's payload