Module Tezos_raw_protocol_015_PtLimaPt.Ticket_costs

This module contains constants and utility functions for gas metering functions used for extracting and handling tickets for the global ticket balance table.

module Constants : sig ... end

val consume_gas_steps : Alpha_context.t -> step_cost:Alpha_context.Gas.cost -> num_steps:int -> Alpha_context.t Tezos_protocol_environment_015_PtLimaPt.Error_monad.tzresult

consume_gas_steps ctxt ~num_steps consumes gas corresponding to a given num_steps and step_cost. It's useful for paying for gas upfront where the number of steps can be determined.

This function is generic and should probably be moved. See issue https://gitlab.com/tezos/tezos/-/issues/1950.

val has_tickets_of_ty_cost : ('a, _) Script_typed_ir.ty -> Saturation_repr.may_saturate Saturation_repr.t

has_tickets_of_ty_cost ty returns the cost of producing a has_tickets, used internally in the Ticket_scanner module.

val negate_cost : Tezos_protocol_environment_015_PtLimaPt.Z.t -> Alpha_context.Gas.cost

negate_cost z returns the cost of negating the given value z.

val add_int_cost : Script_int.n Script_int.num -> Script_int.n Script_int.num -> Alpha_context.Gas.cost

add_int_cost n1 n2 returns the cost of adding the values n1 and n2.

val add_z_cost : Tezos_protocol_environment_015_PtLimaPt.Z.t -> Tezos_protocol_environment_015_PtLimaPt.Z.t -> Alpha_context.Gas.cost

add_z_cost z1 z2 returns the cost of adding the values z1 and z2.