Module Tezos_raw_protocol_015_PtLimaPt.Ticket_token

A module for handling ticket-tokens. A ticket-token represents the combination of a ticketer (creator of a ticket) and the content. That is, a ticket comprises a ticket-token and an amount.

type ex_token =

| Ex_token : {

ticketer : Alpha_context.Contract.t;

contents_type : 'a Script_typed_ir.comparable_ty;

contents : 'a;

} -> ex_token

A type for representing existentially quantified ticket-tokens. A ticket-token consists of a pair of ticketer and contents.

val token_and_amount_of_ex_ticket : Ticket_scanner.ex_ticket -> ex_token * Script_typed_ir.ticket_amount

token_and_amount_of_ex_ticket ex_ticket returns the token and amount of the given ticket ex_ticket.