Module Tezos_raw_protocol_015_PtLimaPt.Tx_rollup_l2_address

This module introduces the types used to identify ticket holders within a transaction rollup.

The hash of a BLS public key is used as the primary identifier of ticket holders within a transaction rollup.

include module type of Tezos_protocol_environment_015_PtLimaPt.Bls.Public_key_hash with type t = Tezos_protocol_environment_015_PtLimaPt.Bls.Public_key_hash.t

type t = Tezos_protocol_environment_015_PtLimaPt.Bls.Public_key_hash.t

val pp : Tezos_protocol_environment_015_PtLimaPt.Format.formatter -> t -> unit

val pp_short : Tezos_protocol_environment_015_PtLimaPt.Format.formatter -> t -> unit

val (=) : t -> t -> bool

val (<>) : t -> t -> bool

val (<) : t -> t -> bool

val (<=) : t -> t -> bool

val (>=) : t -> t -> bool

val (>) : t -> t -> bool

val compare : t -> t -> int

val equal : t -> t -> bool

val max : t -> t -> t

val min : t -> t -> t

val to_bytes : t -> bytes

val of_bytes_opt : bytes -> t option

val of_bytes_exn : bytes -> t

val to_b58check : t -> string

val to_short_b58check : t -> string

val of_b58check_exn : string -> t

val of_b58check_opt : string -> t option

type Tezos_protocol_environment_015_PtLimaPt.Base58.data +=

| Data of t

val b58check_encoding : t Tezos_protocol_environment_015_PtLimaPt.Base58.encoding

val encoding : t Tezos_protocol_environment_015_PtLimaPt.Data_encoding.t

val rpc_arg : t Tezos_protocol_environment_015_PtLimaPt.RPC_arg.t

module Set : sig ... end

module Map : sig ... end

val zero : t

type address = t

val in_memory_size : t -> Cache_memory_helpers.sint

in_memory_size a returns the number of bytes allocated in RAM for a.

val size : t -> int

size a returns the number of bytes allocated in an inbox to store a.

module Indexable : sig ... end