Module Context.Address_metadata

The metadata associated to an address.

val get : t -> Tx_rollup_l2_context_sig.address_index -> Tx_rollup_l2_context_sig.metadata option m

get ctxt idx returns the current metadata associated to the address indexed by idx.

val incr_counter : t -> Tx_rollup_l2_context_sig.address_index -> t m

incr_counter ctxt idx increments the counter of the address indexed by idx.

This function can fail with Counter_overflow iff the counter has reached the Int64.max_int limit.

This function can fail with Unknown_address_index if idx has not been associated with a layer-2 address already.

val init_with_public_key : t -> Tx_rollup_l2_context_sig.address_index -> Tezos_protocol_environment_015_PtLimaPt.Bls.Public_key.t -> t m

init_with_public_key ctxt idx pk initializes the metadata associated to the address indexed by idx.

This can fails with Metadata_already_initialized if this function has already been called with idx.