Module Tx_rollup_l2_apply.Message_result

type transaction_result =

| Transaction_success

| Transaction_failure of {

index : int;

reason : Tezos_protocol_environment_015_PtLimaPt.Error_monad.error;

}

A transaction inside a batch can either be a success or a failure.

In the case of a failure, we store the operation's index which failed with the reason it failed.

type deposit_result =

| Deposit_success of indexes

| Deposit_failure of Tezos_protocol_environment_015_PtLimaPt.Error_monad.error

A deposit can either be a success or a failure. The created indexes must are kept only when the deposit is a success. In the other case, we return the reason why the deposit failed.

The operations are versioned (see Tx_rollup_l2_batch). Therefore, we introduce the operation results for each version.

module Batch_V1 : sig ... end

type message_result =

| Deposit_result of deposit_result

| Batch_V1_result of Batch_V1.t

type t = message_result * Alpha_context.Tx_rollup_withdraw.t list

val encoding : t Tezos_protocol_environment_015_PtLimaPt.Data_encoding.t