Module Tezos_raw_protocol_015_PtLimaPt.Validate_errors

type operation_conflict =

| Operation_conflict of {

existing : Tezos_protocol_environment_015_PtLimaPt.Operation_hash.t;

new_operation : Tezos_protocol_environment_015_PtLimaPt.Operation_hash.t;

}

type used for conflicting operation.

module Consensus : sig ... end

Errors that may arise while validating a consensus operation.

module Voting : sig ... end

Errors that may arise while validating a voting operation.

module Anonymous : sig ... end

Errors that may arise while validating an anonymous operation.

module Manager : sig ... end

Errors that may arise while validating a manager operation.

type Tezos_protocol_environment_015_PtLimaPt.Error_monad.error +=

| Failing_noop_error

module Block : sig ... end