Module Tezos_raw_protocol_015_PtLimaPt.Zk_rollup_scalar

Representation of scalars used by the ZK Rollup alongside manipulation functions

type t = Tezos_protocol_environment_015_PtLimaPt.Bls.Primitive.Fr.t

Scalars are transparently BLS12-381 scalars

val of_z : Tezos_protocol_environment_015_PtLimaPt.Z.t -> t

Safe conversion from Z.t. If the numerical value is not in the field, modulo reduction is applied.

val of_bits : string -> t

Safe conversion from bits, represented as a string. If the numerical value is not in the field, modulo reduction is applied.