Module Tezos_raw_protocol_015_PtLimaPt.Zk_rollup_state_repr

type t = Zk_rollup_scalar.t array

The state of a ZK Rollup is an opaque array of scalars, and represents the L1's view of the L2 state. Although the length of this array is unbound, this type should describe a succinct representation of the entire RU state. Upon origination, the length of a ZKRU's state is fixed.

val encoding : t Tezos_protocol_environment_015_PtLimaPt.Data_encoding.t